requirements/common.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
django-formtools==2.2
django-guardian==2.3.0
django-import-export==2.5.0
django-oauth-toolkit==1.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.19.0
idna==2.10
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
